assignment2/Part2/requirements.txt
aws-cdk.aws-s3
aws-cdk.aws-dynamodb
aws-cdk.aws-lambda
aws-cdk-lib
constructs
boto3